Feat Reduce bundle size NPM
